Rewritten Mac font size code.
A little introduction is necessary to describe the crux of the problem that this commit is trying to solve. The unit terminology is all messed up in the first place and the documentation only makes it more confusing:
- HTML defines "px" unit as 1/96 inch and "pt" unit as 1/72 inch (https://www.w3schools.com/cssref/css_units.asp)
- System.Drawing on Windows uses "pixel" for real device pixels (could be anything between 1/96 inch to 1/480 inch, depending on Windows version and display settings) and "point" for 1/72 inch.
- CoreGraphics uses "point" for a scaled unit that doesn't correspond to physical pixel, neither does it correspond to typographical point (1/72 inch). In context of computer displays the CoreGraphics point is 1/96 inch. In context of PDF or printers it is 1/72 inch.
- Our System.Drawing (further referred as SD) implementation on CoreGraphics (further referred as CG) maps CG point to SD pixel. This means that to applications the display always looks like a one that has the nominal 96 DPI no matter what the real resolution is.

The font selection in Mac applications is presented in sizes corresponding to CG points, or by definition the HTML "px" and our SD pixels. This introduces the confusing concept that NSFont.PointSize (96 DPI) is not the same thing as System.Drawing.Font.SizeInPoints (72 DPI). At the same time it is also true that NSFont.PointSize is the same thing as System.Drawing.Font.Size as long as System.Drawing.Font.Unit == GraphicsUnit.Pixel.

In this patch we try to preserve the size in System.Drawing.Font in the unit that is more natural to the platform. Thus for system fonts on Mac the unit is changed to GraphicsUnit.Pixel (CG point), while on Windows it will remain GraphicsUnit.Point. Similarly, all UI font selectors on Mac now use SD pixel units (CG points).