Mac: Allow closing formMain without quitting the application.
Related fixes and changes:
Fixed synchronization of Control.is_visible flag and the real visibility of the window.
The parent argument in ShowDialog() for Settings, Accounts and FormOAuthAuthentication set to null, to allow showing the forms when the main form is hidden.
Added centering modal dialogs on the screen if their parent is null.
Fixed treating the "Application.MessageLoop" property in Application.RunLoop.
Added "Close" menu item into the File menu of formMain. Added "Main Form" menu item into the "Window" menu.
Added showing and activating the formMain when re-opening the application (by clicking the app icon in the dock or in the Launcher, by double-clicking the app icon in the Finder, etc).